+++ src/utils/authentication/AuthService.js
@@ -13,7 +13,7 @@ const hackauthoritykey = "oidc.hack.authority";
 
 const pathKey = "powsybl-study-app-current-path";
 
-function initializeAuthentication(dispatch, isSilentRenew) {
+function initializeAuthentication(dispatch, isSilentRenew, idpSettings) {
 
     if (process.env.REACT_APP_USE_AUTHENTICATION === "false") {
         let userManager = new UserManagerMock({});
@@ -22,7 +22,7 @@ function initializeAuthentication(dispatch, isSilentRenew) {
         }
         return  Promise.resolve(userManager);
     } else {
-        return fetch('idpSettings.json')
+        return idpSettings
             .then(r => r.json())
             .then(idpSettings => {
                 /* hack to ignore the iss check. XXX TODO to remove */
